requirements.txt
amqp==2.6.1
appdirs==1.4.3
arabic-reshaper==2.1.1
asgiref==3.2.10
autopep8==1.4.4
beautifulsoup4==4.9.1
billiard==3.6.3.0
boto3==1.5.24
botocore==1.8.50
braintree==4.6.0
CacheControl==0.12.6
celery==4.4.2
certifi==2019.9.11
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.6
dateparser==0.7.2
defusedxml==0.6.0
dictop==0.2.1
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
Django==2.2.16
django-admin-autocomplete-filter==0.7.1
django-admin-global-sidebar==0.2.0
django-admin-honeypot==1.1.0
django-admin-list-filter-dropdown==1.0.3
django-admin-tools==0.9.0
django-allauth==0.44.0
django-author==1.2.0
django-autoslug==1.9.7
django-bootstrap4==2.3.1
django-braces==1.14.0
django-cacheops==6.1
django-ckeditor==5.9.0
django-ckeditor-5==0.2.4
django-cors-headers==3.7.0
django-countries==7.0
django-crispy-forms==1.10.0
django-debug-toolbar==2.1
django-discover-runner==1.0
django-environ==0.4.5
django-extensions==3.1.1
django-file-form==3.2.2
django-filter==2.4.0
django-horizontal-list-filter==0.2.7
django-import-export==2.5.0
django-import-export-celery==1.3
django-js-asset==1.2.2
django-mathfilters==0.4.0
django-middleware-global-request==0.3.1
django-model-utils==4.1.1
django-mptt==0.11.0
django-mysql==3.11.1
django-rename-app==0.1.2
django-role-permissions==2.2.1
django-simple-tags==0.6.0
django-social-share==2.2.1
django-static-fontawesome==5.15.4.0
django-static-jquery3==5.1.0
django-storages==1.7.1
django-tables2==2.3.4
django-taggit==1.3.0
django-timezone-field==4.1.1
django-tinymce==3.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
docutils==0.19
et-xmlfile==1.0.1
fake-factory==9999.9.9
Faker==2.0.4
future==0.18.2
gevent==21.12.0
greenlet==1.1.3
gunicorn==19.9.0
html5lib==1.1
idna==3.0
ipaddr==2.2.0
jdcal==1.4.1
jet-bridge-base==0.9.4
jet-django==1.0.7
jmespath==0.10.0
kombu==4.6.11
lockfile==0.12.2
magic-import==0.2.0
Markdown==3.2.2
MarkupPy==1.14
msgpack==0.6.2
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.6
packaging==20.3
pep517==0.8.2
Pillow==9.0.1
progress==1.5
prompt-toolkit==2.0.9
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.5.0
pycparser==2.20
PyJWT==2.4.0
pyparsing==2.4.6
PyPDF2==1.27.5
python-bidi==0.4.2
python-crontab==2.5.1
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.2.0
pytoml==0.1.21
pytz==2019.3
PyYAML==6.0
redis==3.5.3
regex==2022.8.17
reportlab==3.6.11
requests==2.28.1
requests-oauthlib==1.3.0
retrying==1.3.3
s3transfer==0.1.13
sentry-sdk==0.20.3
setuptools-git==1.2
setuptools-scm==3.3.3
six==1.13.0
social-auth-core==4.0.3
soupsieve==2.2
SQLAlchemy==1.3.11
sqlparse==0.3.0
sslcommerz-lib==1.0
tablib==3.0.0
text-unidecode==1.3
toml==0.10.2
tzdata==2022.2
tzlocal==2.0.0
urllib3==1.26.5
vine==1.3.0
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zope.event==4.5.0
zope.interface==5.4.0
py3-validate-email==1.0.5.post1
django-admin-sortable2==1.0.2
django-admin-list-filter-dropdown==1.0.3